requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
cycler==0.11.0
Django
fonttools
h11==0.13.0
idna==3.3
iieba
kiwisolver==1.4.4
matplotlib
numpy
outcome==1.2.0
packaging==21.3
pandas
Pillow==9.2.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.11
wordcloud
wsproto==1.1.0